src/requirements.txt
python-dotenv
bullmq
docling
asyncio